Refactor config and folder structure
