+++ .gitignore
@@ -13,6 +13,9 @@ bin/*
 # Output of the go coverage tool, specifically when used with LiteIDE
 *.out
 
+# ignore .env
+.env
+
 ### Go Patch ###
 /vendor/
 /Godeps/
