+++ go.mod
@@ -3,7 +3,8 @@ module github.com/maria-robobug/animal-api
 go 1.12
 
 require (
-	github.com/gorilla/mux v1.7.0
-	github.com/sirupsen/logrus v1.4.0
+	github.com/caarlos0/env v3.5.0+incompatible
+	github.com/gorilla/mux v1.7.1
+	github.com/joho/godotenv v1.3.0
 	github.com/stretchr/testify v1.3.0
 )
