+++ internal/rest/dog/handler_test.go
@@ -1,4 +1,4 @@
-package rest_test
+package dog_test
 
 import (
 	"encoding/json"
@@ -10,9 +10,9 @@ import (
 	"os"
 	"testing"
 
-	"github.com/maria-robobug/animal-api/pkg/mock"
+	"github.com/maria-robobug/animal-api/internal/mock"
 
-	"github.com/maria-robobug/animal-api/pkg/dog/rest"
+	"github.com/maria-robobug/animal-api/internal/rest/dog"
 	"github.com/stretchr/testify/assert"
 )
