+++ internal/rest/dog/server.go
@@ -1,4 +1,4 @@
-package rest
+package dog
 
 import (
 	"errors"
